requirements.txt
# 基础依赖
pytest>=7.0.0
requests>=2.28.0

# 测试报告相关
pytest-html>=3.1.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
allure-pytest>=2.12.0

# 日志和工具
PyYAML>=6.0
python-dotenv>=0.19.0